Removing old version code
