+++ server/src/main/scala/za/co/absa/atum/server/api/controller/FlowController.scala
@@ -16,23 +16,19 @@
 
 package za.co.absa.atum.server.api.controller
 
-import za.co.absa.atum.model.dto.{CheckpointDTO, CheckpointQueryDTO, CheckpointV2DTO}
+import za.co.absa.atum.model.dto.CheckpointV2DTO
 import za.co.absa.atum.server.model.ErrorResponse
-import za.co.absa.atum.server.model.SuccessResponse.{MultiSuccessResponse, PaginatedResponse}
+import za.co.absa.atum.server.model.SuccessResponse.PaginatedResponse
 import zio.IO
 import zio.macros.accessible
 
 @accessible
 trait FlowController {
   def getFlowCheckpointsV2(
-    checkpointQueryDTO: CheckpointQueryDTO
-  ): IO[ErrorResponse, MultiSuccessResponse[CheckpointDTO]]
-
-  def getFlowCheckpoints(
-                          flowId: Long,
-                          limit: Option[Int],
-                          offset: Option[Long],
-                          checkpointName: Option[String],
+    flowId: Long,
+    limit: Option[Int],
+    offset: Option[Long],
+    checkpointName: Option[String],
   ): IO[ErrorResponse, PaginatedResponse[CheckpointV2DTO]]
 
 }
