+++ server/src/main/scala/za/co/absa/atum/server/api/service/FlowServiceImpl.scala
@@ -18,28 +18,13 @@ package za.co.absa.atum.server.api.service
 
 import za.co.absa.atum.model.dto._
 import za.co.absa.atum.server.api.exception.ServiceError
-import za.co.absa.atum.server.api.exception.ServiceError._
 import za.co.absa.atum.server.api.repository.FlowRepository
-import za.co.absa.atum.server.model.{CheckpointFromDB, PaginatedResult}
+import za.co.absa.atum.server.model.PaginatedResult
 import zio._
 
 class FlowServiceImpl(flowRepository: FlowRepository) extends FlowService with BaseService {
 
-  override def getFlowCheckpoints(checkpointQueryDTO: CheckpointQueryDTO): IO[ServiceError, Seq[CheckpointDTO]] = {
-    for {
-      checkpointsFromDB <- repositoryCall(
-        flowRepository.getFlowCheckpoints(checkpointQueryDTO),
-        "getFlowCheckpoints"
-      )
-      checkpointDTOs <- ZIO.foreach(checkpointsFromDB) { checkpointFromDB =>
-        ZIO
-          .fromEither(CheckpointFromDB.toCheckpointDTO(checkpointQueryDTO.partitioning, checkpointFromDB))
-          .mapError(error => GeneralServiceError(error.getMessage))
-      }
-    } yield checkpointDTOs
-  }
-
-  override def getFlowCheckpointsV2(
+ override def getFlowCheckpointsV2(
      partitioningId: Long,
      limit: Option[Int],
      offset: Option[Long],
@@ -50,6 +35,7 @@ class FlowServiceImpl(flowRepository: FlowRepository) extends FlowService with B
       "getFlowCheckpointsV2"
     )
   }
+
 }
 
 object FlowServiceImpl {
