+++ server/src/test/scala/za/co/absa/atum/server/api/controller/FlowControllerUnitTests.scala
@@ -20,7 +20,7 @@ import org.mockito.Mockito.{mock, when}
 import za.co.absa.atum.server.api.TestData
 import za.co.absa.atum.server.api.exception.ServiceError._
 import za.co.absa.atum.server.api.service.FlowService
-import za.co.absa.atum.server.model.{InternalServerErrorResponse, NotFoundErrorResponse, Pagination}
+import za.co.absa.atum.server.model.{NotFoundErrorResponse, Pagination}
 import za.co.absa.atum.server.model.PaginatedResult.{ResultHasMore, ResultNoMore}
 import zio._
 import zio.test.Assertion.failsWithA
@@ -28,11 +28,6 @@ import zio.test._
 
 object FlowControllerUnitTests extends ZIOSpecDefault with TestData {
   private val flowServiceMock = mock(classOf[FlowService])
-
-  when(flowServiceMock.getFlowCheckpoints(checkpointQueryDTO1))
-    .thenReturn(ZIO.fail(GeneralServiceError("boom!")))
-  when(flowServiceMock.getFlowCheckpoints(checkpointQueryDTO2))
-    .thenReturn(ZIO.succeed(Seq(checkpointDTO2)))
 
   when(flowServiceMock.getFlowCheckpointsV2(1L, Some(5), Some(2), None))
     .thenReturn(ZIO.succeed(ResultHasMore(Seq(checkpointV2DTO1))))
@@ -45,31 +40,19 @@ object FlowControllerUnitTests extends ZIOSpecDefault with TestData {
 
   override def spec: Spec[TestEnvironment with Scope, Any] = {
     suite("FlowControllerSuite")(
-      suite("GetFlowCheckpointsSuite")(
-        test("Returns expected InternalServerErrorResponse") {
-          assertZIO(FlowController.getFlowCheckpointsV2(checkpointQueryDTO1).exit)(
-            failsWithA[InternalServerErrorResponse]
-          )
-        },
-        test("Returns expected CheckpointDTO") {
-          for {
-            result <- FlowController.getFlowCheckpointsV2(checkpointQueryDTO2)
-          } yield assertTrue(result.data == Seq(checkpointDTO2))
-        }
-      ),
       suite("GetFlowCheckpointsV2Suite")(
         test("Returns expected Seq[CheckpointV2DTO] with Pagination indicating there is more data available") {
           for {
-            result <- FlowController.getFlowCheckpoints(1L, Some(5), Some(2), None)
+            result <- FlowController.getFlowCheckpointsV2(1L, Some(5), Some(2), None)
           } yield assertTrue(result.data == Seq(checkpointV2DTO1) && result.pagination == Pagination(5, 2, hasMore = true))
         },
         test("Returns expected Seq[CheckpointV2DTO] with Pagination indicating there is no more data available") {
           for {
-            result <- FlowController.getFlowCheckpoints(2L, Some(5), Some(0), None)
+            result <- FlowController.getFlowCheckpointsV2(2L, Some(5), Some(0), None)
           } yield assertTrue(result.data == Seq(checkpointV2DTO2) && result.pagination == Pagination(5, 0, hasMore = false))
         },
         test("Returns expected NotFoundServiceError when service returns NotFoundServiceError") {
-          assertZIO(FlowController.getFlowCheckpoints(3L, Some(5), Some(0), None).exit)(
+          assertZIO(FlowController.getFlowCheckpointsV2(3L, Some(5), Some(0), None).exit)(
             failsWithA[NotFoundErrorResponse]
           )
         }
