+++ server/src/test/scala/za/co/absa/atum/server/api/repository/FlowRepositoryUnitTests.scala
@@ -19,7 +19,7 @@ package za.co.absa.atum.server.api.repository
 import org.mockito.Mockito.{mock, when}
 import za.co.absa.atum.server.api.TestData
 import za.co.absa.atum.server.api.database.flows.functions.GetFlowCheckpointsV2.GetFlowCheckpointsArgs
-import za.co.absa.atum.server.api.database.flows.functions.{GetFlowCheckpoints, GetFlowCheckpointsV2}
+import za.co.absa.atum.server.api.database.flows.functions.GetFlowCheckpointsV2
 import za.co.absa.atum.server.api.exception.DatabaseError
 import za.co.absa.db.fadb.exceptions.DataNotFoundException
 import za.co.absa.atum.server.model.PaginatedResult.ResultNoMore
@@ -30,18 +30,6 @@ import zio.test._
 import za.co.absa.db.fadb.status.{FunctionStatus, Row}
 
 object FlowRepositoryUnitTests extends ZIOSpecDefault with TestData {
-
-  private val getFlowCheckpointsMock = mock(classOf[GetFlowCheckpoints])
-
-  when(getFlowCheckpointsMock.apply(checkpointQueryDTO1)).thenReturn(ZIO.fail(new Exception("boom!")))
-  when(getFlowCheckpointsMock.apply(checkpointQueryDTO2))
-    .thenReturn(
-      ZIO.right(
-        Seq(Row(FunctionStatus(0, "success"), checkpointFromDB1), Row(FunctionStatus(0, "success"), checkpointFromDB2))
-      )
-    )
-
-  private val getFlowCheckpointsMockLayer = ZLayer.succeed(getFlowCheckpointsMock)
 
   private val getFlowCheckpointsV2Mock = mock(classOf[GetFlowCheckpointsV2])
 
@@ -62,18 +50,6 @@ object FlowRepositoryUnitTests extends ZIOSpecDefault with TestData {
   override def spec: Spec[TestEnvironment with Scope, Any] = {
 
     suite("FlowRepositoryIntegrationSuite")(
-      suite("GetFlowCheckpointsSuite")(
-        test("Returns expected DatabaseError") {
-          assertZIO(FlowRepository.getFlowCheckpoints(checkpointQueryDTO1).exit)(
-            failsWithA[DatabaseError]
-          )
-        },
-        test("Returns expected Left with StatusException") {
-          for {
-            result <- FlowRepository.getFlowCheckpoints(checkpointQueryDTO2)
-          } yield assertTrue(result == Seq(checkpointFromDB1, checkpointFromDB2))
-        }
-      ),
       suite("GetFlowCheckpointsV2Suite")(
         test("Returns expected Right with CheckpointV2DTO") {
           for {
@@ -88,7 +64,6 @@ object FlowRepositoryUnitTests extends ZIOSpecDefault with TestData {
       )
     ).provide(
       FlowRepositoryImpl.layer,
-      getFlowCheckpointsMockLayer,
       getFlowCheckpointsV2MockLayer
     )
