+++ server/src/test/scala/za/co/absa/atum/server/api/service/FlowServiceUnitTests.scala
@@ -18,8 +18,6 @@ package za.co.absa.atum.server.api.service
 
 import org.mockito.Mockito.{mock, when}
 import za.co.absa.atum.server.api.TestData
-import za.co.absa.atum.server.api.exception.DatabaseError.GeneralDatabaseError
-import za.co.absa.atum.server.api.exception.ServiceError
 import za.co.absa.atum.server.api.repository.FlowRepository
 import za.co.absa.atum.server.model.PaginatedResult.ResultHasMore
 import za.co.absa.atum.server.api.exception.DatabaseError.NotFoundDatabaseError
@@ -31,10 +29,6 @@ import zio.test._
 object FlowServiceUnitTests extends ZIOSpecDefault with TestData {
   private val flowRepositoryMock = mock(classOf[FlowRepository])
 
-  when(flowRepositoryMock.getFlowCheckpoints(checkpointQueryDTO1)).thenReturn(ZIO.fail(GeneralDatabaseError("boom!")))
-  when(flowRepositoryMock.getFlowCheckpoints(checkpointQueryDTO2))
-    .thenReturn(ZIO.succeed(Seq(checkpointFromDB2)))
-
   when(flowRepositoryMock.getFlowCheckpointsV2(1L, None, None, None))
     .thenReturn(ZIO.succeed(ResultHasMore(Seq(checkpointV2DTO1))))
   when(flowRepositoryMock.getFlowCheckpointsV2(2L, None, None, None))
@@ -45,20 +39,6 @@ object FlowServiceUnitTests extends ZIOSpecDefault with TestData {
   override def spec: Spec[TestEnvironment with Scope, Any] = {
 
     suite("FlowServiceSuite")(
-      suite("GetFlowCheckpointsSuite")(
-        test("Returns expected ServiceError") {
-          assertZIO(FlowService.getFlowCheckpoints(checkpointQueryDTO1).exit)(
-            failsWithA[ServiceError]
-          )
-        },
-        test("Returns expected Seq[CheckpointDTO]") {
-          for {
-            result <- FlowService.getFlowCheckpoints(checkpointQueryDTO2)
-          } yield assertTrue {
-            result == Seq(checkpointDTO2)
-          }
-        }
-      ),
       suite("GetFlowCheckpointsV2Suite")(
         test("Returns expected PaginatedResult[CheckpointV2DTO]") {
           for {
